moved SensorChoice
